Resolve #318: Remove `sendBroadcastAction` method.
In fact there is no need in this method anymore and I would call it
unsafe for usage.

It was initially created for managing state changes across several
stores but now we can specify dependencies for a store and send actions
to other stores using the `sendAction` method by a store name.